project/requirements.txt
Django
psycopg2
djangorestframework
django-filter
requests
beautifulsoup4
lxml
